Bug fix in the MST creation algorithm

+++ tests/quality/test_duplicates.py
@@ -116,7 +116,7 @@ class TestDuplicates:
                 "child_size": np.array([]),
             },
             "membership_strengths": np.array([], dtype=np.float32),
-            "k_neighbors": np.array([], dtype=np.int32),
+            "k_neighbors": np.array([], dtype=np.int64),
             "k_distances": np.array([], dtype=np.float32),
         }
 
@@ -143,7 +143,7 @@ class TestDuplicates:
                 "child_size": np.array([]),
             },
             "membership_strengths": np.array([], dtype=np.float32),
-            "k_neighbors": np.array([], dtype=np.int32),
+            "k_neighbors": np.array([], dtype=np.int64),
             "k_distances": np.array([], dtype=np.float32),
         }
 
@@ -172,7 +172,7 @@ class TestDuplicates:
                 "child_size": np.array([]),
             },
             "membership_strengths": np.array([], dtype=np.float32),
-            "k_neighbors": np.array([], dtype=np.int32),
+            "k_neighbors": np.array([], dtype=np.int64),
             "k_distances": np.array([], dtype=np.float32),
         }
